cleanup: use upstream accounts/keystore
