--- accounts/keystore/testdata/v1_test_vector.json
@@ -1,28 +1,0 @@
-{
-    "test1": {
-        "json": {
-            "Crypto": {
-                "cipher": "aes-128-cbc",
-                "cipherparams": {
-                    "iv": "35337770fc2117994ecdcad026bccff4"
-                },
-                "ciphertext": "6143d3192db8b66eabd693d9c4e414dcfaee52abda451af79ccf474dafb35f1bfc7ea013aa9d2ee35969a1a2e8d752d0",
-                "kdf": "scrypt",
-                "kdfparams": {
-                    "dklen": 32,
-                    "n": 262144,
-                    "p": 1,
-                    "r": 8,
-                    "salt": "9afcddebca541253a2f4053391c673ff9fe23097cd8555d149d929e4ccf1257f"
-                },
-                "mac": "3f3d5af884b17a100b0b3232c0636c230a54dc2ac8d986227219b0dd89197644",
-                "version": "1"
-            },
-            "address": "cb61d5a9c4896fb9658090b597ef0e7be6f7b67e",
-            "id": "e25f7c1f-d318-4f29-b62c-687190d4d299",
-            "version": "1"
-        },
-        "password": "g",
-        "priv": "d1b1178d3529626a1a93e073f65028370d14c7eb0936eb42abef05db6f37ad7d"
-    }
-}
