+++ scripts/geth-allowed-packages.txt
@@ -1,5 +1,6 @@
 "github.com/ethereum/go-ethereum/accounts"
 "github.com/ethereum/go-ethereum/accounts/external"
+"github.com/ethereum/go-ethereum/accounts/keystore"
 "github.com/ethereum/go-ethereum/accounts/scwallet"
 "github.com/ethereum/go-ethereum/common"
 "github.com/ethereum/go-ethereum/common/bitutil"
